Nomal Mapping For TA

+++ Code/OysterGraphics/Shader/Passes/Post/PostPass.hlsl
@@ -17,7 +17,7 @@ void main( uint3 DTid : SV_DispatchThreadID )
 	float4 Light = Diffuse[DTid.xy] + saturate(Specular[DTid.xy]);
 	float4 Amb = float4(Ambient[DTid.xy/2].xyz * Ambient[DTid.xy/2].w, 0);
 	float4 GUI;
-	uint2 index = DTid.xy/2 + uint2(Pixels.x/2,0);
+	uint2 index = DTid.xy/2 + uint2((uint)Pixels.x/(uint)2,0);
 	float3 PostLight = Amb.xyz * AmbFactor;
 	PostLight = PostLight + Light.xyz;
 	GUI = float4(Ambient[index]);
